requirements.txt
django==3.2.7
djangorestframework==3.12.4
psycopg2==2.9.1
python-dotenv==0.19.0
pytest-black==0.3.12



